refactor(frontend): split playtest update flow
