+++ engine/examples/editor-playtest/src/playtest_update.rs
@@ -1,0 +1,282 @@
+use super::*;
+
+impl Playtest {
+    pub(super) fn init_gameplay(&mut self) {
+        self.shadow_material = upload_shadow_texture();
+        self.particle_material = upload_particle_texture();
+
+        // Empty manifest? Boot to a clear-coloured screen.
+        if ROOMS.is_empty() {
+            return;
+        };
+
+        // Player init: prefer PLAYER_CONTROLLER (cook output)
+        // for spawn + character; fall back to the bare
+        // PLAYER_SPAWN for placeholder manifests. The spawn room
+        // may be a manual portal room rather than room zero.
+        let (spawn, character) = match PLAYER_CONTROLLER {
+            Some(pc) => {
+                let character = CHARACTERS
+                    .get(pc.character.to_usize())
+                    .map(RuntimeCharacter::from_record);
+                (pc.spawn, character)
+            }
+            None => (PLAYER_SPAWN, None),
+        };
+        if ROOMS.get(spawn.room.to_usize()).is_none() {
+            return;
+        };
+        self.load_runtime_models();
+        self.rebuild_box_prop_runtime();
+        self.spawn = RoomPoint::new(spawn.x, spawn.y, spawn.z);
+        self.character = character;
+        self.motor
+            .snap_to(self.spawn, Angle::from_q12(spawn.yaw as u16));
+        self.room_index = spawn.room;
+        self.anim_state = PlayerAnim::Idle;
+        self.anim_start_tick = SimTick::ZERO;
+        self.anim_lock_until_tick = SimTick::ZERO;
+        self.active_interactable = None;
+        self.checkpoint = None;
+        self.message_overlay = None;
+        self.box_prop_broken = [0; BOX_PROP_BROKEN_WORDS];
+        self.box_prop_fall = [BoxPropFallState::EMPTY; MAX_BOX_PROP_STATE];
+        self.box_prop_break_events = [BoxPropBreakEvent::EMPTY; MAX_BOX_PROP_BREAK_EVENTS];
+        self.camera.snap_to_player_with_yaw(
+            self.camera_target(None, false),
+            self.camera_config(),
+            CAMERA_START_YAW,
+        );
+        self.render_camera = world_camera_from_position_focus(
+            PROJECTION,
+            self.camera.position(),
+            self.camera.focus(),
+        );
+        #[cfg(feature = "cd-stream-bench")]
+        self.bootstrap_streamed_room_window();
+        #[cfg(not(feature = "cd-stream-bench"))]
+        self.load_active_room_window();
+        #[cfg(feature = "cd-stream-benchmark")]
+        cd_stream::run_benchmark();
+    }
+
+    pub(super) fn update_gameplay(&mut self, ctx: &mut Ctx) {
+        self.portal_debug_log_cooldown = self.portal_debug_log_cooldown.saturating_sub(1);
+        self.step_streaming_jobs(ctx);
+
+        if ctx.just_pressed(button::R3) {
+            self.lock_target = match self.lock_target {
+                Some(_) => None,
+                None => self.find_best_lock_target(LOCK_RANGE),
+            };
+            self.lock_switch_stick_held = false;
+            self.soft_lock_target = None;
+        }
+        if ctx.just_pressed(COLLISION_DEBUG_BUTTON) {
+            self.show_collision_debug = !self.show_collision_debug;
+        }
+
+        if self.message_overlay.is_some() {
+            if ctx.just_pressed(INTERACT_BUTTON) || ctx.just_pressed(button::CIRCLE) {
+                self.message_overlay = None;
+            }
+            self.camera_turning_last_tick = false;
+            return;
+        }
+
+        if !ctx.pad.is_analog() {
+            self.camera_turning_last_tick = false;
+            return;
+        }
+
+        if ctx.just_pressed(button::SELECT) {
+            self.free_orbit = !self.free_orbit;
+        }
+        let delta_vblanks = 1u16;
+        self.advance_box_prop_break_events(delta_vblanks);
+        self.advance_box_prop_falls(delta_vblanks);
+        if CAMERA_SWEEP_ENABLED {
+            self.update_camera_sweep(delta_vblanks);
+            return;
+        }
+        if self.free_orbit {
+            let (right_x, right_y) = ctx.pad.sticks.right_centered();
+            self.camera_turning_last_tick = abs_i16(right_x) >= CAMERA_STICK_DEADZONE;
+            self.orbit_yaw = self.orbit_yaw.add_signed_q12(scale_i16_by_vblanks(
+                stick_to_yaw_delta(psx_engine::InputAxis::new(right_x.saturating_neg())),
+                delta_vblanks,
+            ));
+            self.orbit_radius = (self.orbit_radius
+                + scale_i32_by_vblanks(
+                    stick_to_radius_delta(psx_engine::InputAxis::new(right_y)),
+                    delta_vblanks,
+                ))
+            .clamp(CAMERA_RADIUS_MIN, CAMERA_RADIUS_MAX);
+            let button_yaw_step =
+                scale_i16_by_vblanks(CAMERA_YAW_STEP.as_q12() as i16, delta_vblanks);
+            let button_radius_step = scale_i32_by_vblanks(CAMERA_RADIUS_STEP, delta_vblanks);
+            if ctx.is_held(button::RIGHT) {
+                self.orbit_yaw = self.orbit_yaw.add_signed_q12(button_yaw_step);
+            }
+            if ctx.is_held(button::LEFT) {
+                self.orbit_yaw = self
+                    .orbit_yaw
+                    .add_signed_q12(button_yaw_step.saturating_neg());
+            }
+            if ctx.is_held(button::UP) {
+                self.orbit_radius = (self.orbit_radius - button_radius_step).max(CAMERA_RADIUS_MIN);
+            }
+            if ctx.is_held(button::DOWN) {
+                self.orbit_radius = (self.orbit_radius + button_radius_step).min(CAMERA_RADIUS_MAX);
+            }
+            self.player_moved_last_tick = false;
+            self.active_interactable = None;
+            telemetry::stage_begin(telemetry::stage::CAMERA);
+            self.render_camera = self.free_orbit_camera();
+            telemetry::stage_end(telemetry::stage::CAMERA);
+            self.refresh_active_room_window_if_needed();
+            #[cfg(all(
+                feature = "world-grid-visible",
+                not(feature = "vis-full-active-chunks")
+            ))]
+            self.prewarm_visible_cell_caches();
+            return;
+        }
+
+        let now = ctx.sim_tick;
+        let action_locked = self.anim_lock_until_tick > now;
+        self.refresh_active_interactable();
+        if !action_locked {
+            if let Some(index) = self.active_interactable {
+                if ctx.just_pressed(INTERACT_BUTTON) && self.activate_interactable(index) {
+                    self.evade_run_hold_ticks = 0;
+                    self.evade_run_hold_consumed = false;
+                    self.camera_turning_last_tick = false;
+                    return;
+                }
+            }
+        }
+        let circle = self.update_evade_run_button(ctx, delta_vblanks);
+        let mut input = if action_locked {
+            CharacterMotorInput::default()
+        } else {
+            motor_input(ctx, self.camera.yaw(), circle.sprint, circle.evade)
+        };
+        if !action_locked && self.motor.action().is_idle() {
+            let started = if ctx.just_pressed(LIGHT_ATTACK_BUTTON) {
+                self.start_player_anim_action(PlayerAnim::LightAttack, now, ctx.video_hz)
+            } else if ctx.just_pressed(HEAVY_ATTACK_BUTTON) {
+                self.start_player_anim_action(PlayerAnim::HeavyAttack, now, ctx.video_hz)
+            } else {
+                false
+            };
+            if started {
+                input = CharacterMotorInput::default();
+            }
+        }
+        let config = self.motor_config();
+        if self.anim_lock_until_tick > now && player_anim_is_attack(self.anim_state) {
+            self.break_box_props_for_attack(config);
+        } else if let Some(trigger) =
+            box_prop_movement_break_trigger(input, config, self.motor.stamina_q12())
+        {
+            self.break_box_props_for_movement(trigger, input, config, delta_vblanks);
+        }
+        telemetry::stage_begin(telemetry::stage::SIM_COLLISION);
+        let mut collision_rooms = [const { CharacterCollisionRoom::EMPTY }; MAX_COLLISION_ROOMS];
+        let collision_room_count = if self.chunked_level() {
+            let catchup = delta_vblanks.min(4) as i32;
+            let margin = config
+                .radius
+                .saturating_add(config.run_speed.saturating_mul(catchup));
+            self.collect_collision_rooms(self.motor.position(), margin, &mut collision_rooms)
+        } else {
+            0
+        };
+        let single_collision_room = if collision_room_count == 1 {
+            collision_rooms[0].room
+        } else {
+            None
+        };
+        let room_collision = match collision_room_count {
+            0 => self
+                .current_collision_room
+                .as_ref()
+                .map(|room| room.collision()),
+            1 => single_collision_room.as_ref().map(|room| room.collision()),
+            _ => None,
+        };
+        let mut blockers = [CharacterCollisionCylinder::EMPTY; MAX_MODEL_INSTANCES];
+        let blocker_count = self.collect_collision_blockers(&mut blockers);
+        let mut aabb_blockers = [CharacterCollisionAabb::EMPTY; MAX_BOX_PROP_BLOCKERS];
+        let aabb_blocker_count = self.collect_box_prop_collision_blockers(&mut aabb_blockers);
+        let collision = if collision_room_count <= 1 {
+            CharacterCollision::new_with_aabbs(
+                room_collision,
+                &blockers[..blocker_count],
+                &aabb_blockers[..aabb_blocker_count],
+            )
+        } else {
+            CharacterCollision::rooms_with_aabbs(
+                &collision_rooms[..collision_room_count],
+                &blockers[..blocker_count],
+                &aabb_blockers[..aabb_blocker_count],
+            )
+        };
+        telemetry::stage_end(telemetry::stage::SIM_COLLISION);
+        telemetry::stage_begin(telemetry::stage::SIM_SOLVE);
+        let motor_frame =
+            self.motor
+                .update_vblanks_with_collision(collision, input, config, delta_vblanks);
+        telemetry::stage_end(telemetry::stage::SIM_SOLVE);
+        self.player_moved_last_tick = motor_frame.moved;
+        telemetry::stage_begin(telemetry::stage::SIM_ROOM_TRACK);
+        if !self.update_current_room_from_player() {
+            self.refresh_active_room_window_if_needed();
+        }
+        telemetry::stage_end(telemetry::stage::SIM_ROOM_TRACK);
+
+        let new_state = if self.anim_lock_until_tick > now {
+            self.anim_state
+        } else {
+            player_anim_from_motor(motor_frame.anim)
+        };
+        if new_state != self.anim_state {
+            self.anim_state = new_state;
+            self.anim_start_tick = now;
+            if new_state.is_motor_fixed_action() {
+                if let Some(character) = self.character {
+                    self.lock_player_anim_action(character, new_state, now, ctx.video_hz);
+                }
+            }
+        }
+
+        if self.lock_target.is_some() {
+            if !self.lock_target_valid(LOCK_BREAK_RANGE) {
+                self.lock_target = None;
+                self.lock_switch_stick_held = false;
+            } else {
+                self.update_lock_target_switch(ctx);
+            }
+        }
+        let (camera_right_x, _) = ctx.pad.sticks.right_centered();
+        self.camera_turning_last_tick =
+            self.lock_target.is_none() && abs_i16(camera_right_x) >= CAMERA_STICK_DEADZONE;
+        if SOFT_LOCK_ENABLED {
+            self.update_soft_lock(ctx);
+        } else {
+            self.soft_lock_target = None;
+            self.soft_lock_suppressed = false;
+        }
+
+        telemetry::stage_begin(telemetry::stage::CAMERA);
+        self.render_camera = self.update_follow_camera(ctx);
+        telemetry::stage_end(telemetry::stage::CAMERA);
+        self.refresh_active_room_window_if_needed();
+        #[cfg(all(
+            feature = "world-grid-visible",
+            not(feature = "vis-full-active-chunks")
+        ))]
+        self.prewarm_visible_cell_caches();
+    }
+}
